flash messages

+++ delete.php
@@ -19,6 +19,16 @@ if ($statement->rowCount() == 0) {
   return;
 }
 
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !==  $_SESSION['user']['id']) {
+  http_response_code(403);
+  echo("HTTP 403 UNAUTHORIZED");
+  return;
+}
+
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
+
+$_SESSION["flash"] = ["message" => "Contact {$contact['name']} deleted."];
 
 header("Location: home.php");
